separate portage cataloger parser impl from cataloger
